+++ packages/smarthr-ui/src/themes/index.ts
@@ -1,2 +1,23 @@
-export * from './tailwind'
-export { createMediaQuery, defaultMediaQuery } from './createMediaQuery'
+export {
+  colors,
+  textColor,
+  fontSize,
+  backgroundColor,
+  lineHeight,
+  boxShadow,
+  spacing,
+  bgColors,
+  paddingBlock,
+  paddingInline,
+  dialogSize,
+} from './tailwind'
+export {
+  createMediaQuery,
+  defaultMediaQuery,
+  type CreatedMediaQueryTheme,
+} from './createMediaQuery'
+export type { AbstractSize, CharRelativeSize } from './createSpacing'
+export { createSpacingByChar, primitiveTokens } from './createSpacing'
+export type { FontSizes } from './createFontSize'
+export { defaultHtmlFontSize } from './createFontSize'
+export { defaultBreakpoint } from './createBreakpoint'